move test module to correct location
