+++ tests/test_train.py
@@ -20,7 +20,7 @@ from rasa.nlu.model import Interpreter
 from rasa.train import train_core, train_nlu, train
 from tests.conftest import DEFAULT_CONFIG_PATH, DEFAULT_NLU_DATA
 from tests.core.conftest import DEFAULT_DOMAIN_PATH_WITH_SLOTS, DEFAULT_STORIES_FILE
-from tests.core.test_model import _fingerprint
+from tests.test_model import _fingerprint
 
 
 @pytest.mark.parametrize(
